add the first working Dockerfile
This isn't great but proves the idea. When built, the resulting image
will be a jupyter notebook tensorflow image with networkx added via pip
and otel-cli slipped in. The entrypoint is modified so otel-cli starts
in server mode in the background and writes incoming traces to files.
These files are visible inside the Jupyter environment and I have gone
no futher yet.